Move to link shortener namespace
