--- Imagizer.Api/Infrastructure/IUrlShortener.cs
@@ -1,6 +1,0 @@
-namespace Imagizer.Api.Infrastructure;
-
-public interface IUrlShortener
-{
-    public Task<string?> ShortenUrl(ShortenerRequestModel shortenerRequest);
-}
