+++ Imagizer.Api/Infrastructure/LinkShortener/IUrlShortener.cs
@@ -1,0 +1,6 @@
+namespace Imagizer.Api.Infrastructure.LinkShortener;
+
+public interface IUrlShortener : IDisposable
+{
+    public Task<string?> ShortenUrl(ShortenerRequestModel shortenerRequest);
+}
