+++ Imagizer.Api/Infrastructure/LinkShortener/ShortenerRequestModel.cs
@@ -1,4 +1,4 @@
-namespace Imagizer.Api.Infrastructure;
+namespace Imagizer.Api.Infrastructure.LinkShortener;
 
 public class ShortenerRequestModel
 {
